Fix up code generator, get `valence` compiling
