+++ crates/valence/src/entity.rs
@@ -1,66 +1,59 @@
-use std::collections::{BTreeMap, HashMap};
-use std::fmt;
-use std::fmt::Formatter;
 use std::num::Wrapping;
-use std::ops::Range;
 
 use bevy_app::{App, CoreSet, Plugin};
 use bevy_ecs::prelude::*;
-use bevy_ecs::query::WorldQuery;
-use glam::{DVec3, UVec3, Vec3};
+use glam::Vec3;
 use rand::rngs::{OsRng, StdRng};
 use rand::{Rng, SeedableRng};
 use rustc_hash::FxHashMap;
-use tracing::{debug, warn};
+use tracing::warn;
 use uuid::Uuid;
-use valence_protocol::byte_angle::ByteAngle;
-use valence_protocol::packet::s2c::play::{
-    EntityAnimationS2c, EntityPositionS2c, EntitySetHeadYawS2c, EntitySpawnS2c,
-    EntityStatusS2c as EntityEventS2c, EntityTrackerUpdateS2c, EntityVelocityUpdateS2c,
-    ExperienceOrbSpawnS2c, MoveRelativeS2c, PlayerSpawnS2c, RotateAndMoveRelativeS2c, RotateS2c,
-};
 pub use valence_protocol::types::Direction;
 use valence_protocol::var_int::VarInt;
 use valence_protocol::{Decode, Encode};
 
-use self::data::TrackedData;
 use crate::client::FlushPacketsSet;
 use crate::component::{
     Despawned, Location, Look, OldLocation, OldPosition, OnGround, Position, UniqueId,
 };
 use crate::config::DEFAULT_TPS;
 use crate::packet::WritePacket;
-use crate::util::{velocity_to_packet_units, Aabb};
-
-pub mod data;
 
 include!(concat!(env!("OUT_DIR"), "/entity_event.rs"));
 include!(concat!(env!("OUT_DIR"), "/entity.rs"));
 
 /// A Minecraft entity's ID according to the protocol.
 ///
-/// IDs should be unique between all spawned entities and should not be modified
-/// after creation. IDs of zero (the default) will be assigned to something else
-/// after the entity is added. If you need to know the ID ahead of time, set
-/// this component to the value returned by [`EntityManager::next_id`] before
-/// spawning.
-#[derive(Component, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Default, Debug)]
+/// IDs should be unique between all spawned entities and should stay constant
+/// during the lifetime of the entity. IDs of -1 (the default) will be
+/// assigned to something else on the tick the entity is added. If you need to
+/// know the ID ahead of time, set this component to the value returned by
+/// [`EntityManager::next_id`] before spawning.
+#[derive(Component, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
 pub struct EntityId(i32);
 
 impl EntityId {
-    pub fn get(&self) -> i32 {
+    /// Returns the underlying entity ID as an integer.
+    pub fn get(self) -> i32 {
         self.0
     }
 }
 
-#[derive(Component, Default)]
+/// Returns an entity ID of -1.
+impl Default for EntityId {
+    fn default() -> Self {
+        Self(-1)
+    }
+}
+
+#[derive(Component, Copy, Clone, PartialEq, Default, Debug)]
 pub struct HeadYaw(pub f32);
 
 /// Entity velocity in m/s.
-#[derive(Component, Copy, Clone, PartialEq, Default, Debug)]
+#[derive(Component, Copy, Clone, Default, Debug)]
 pub struct Velocity(pub Vec3);
 
-#[derive(Component, Default)]
+#[derive(Component, Copy, Clone, Default, Debug)]
 pub struct EntityStatuses(u64);
 
 impl EntityStatuses {
@@ -77,7 +70,7 @@ impl EntityStatuses {
     }
 }
 
-#[derive(Component, Default)]
+#[derive(Component, Default, Debug)]
 pub struct EntityAnimations(u8);
 
 impl EntityAnimations {
@@ -104,15 +97,121 @@ impl EntityAnimations {
 /// - **Falling Block**: Block state
 /// - **Fishing Bobber**: Hook entity ID
 /// - **Warden**: Initial pose
-#[derive(Component, Default)]
+#[derive(Component, Default, Debug)]
 pub struct ObjectData(pub i32);
 
 /// The range of packet bytes for this entity within the cell the entity is
 /// located in. For internal use only.
-#[derive(Component, Default)]
+#[derive(Component, Default, Debug)]
 pub struct PacketByteRange {
     pub(crate) begin: usize,
     pub(crate) end: usize,
+}
+
+/// Cache for all the tracked data of an entity. Used for the
+/// [`EntityTrackerUpdateS2c`][packet] packet.
+///
+/// [packet]: valence_protocol::packet::s2c::play::EntityTrackerUpdateS2c
+#[derive(Component, Default, Debug)]
+pub struct TrackedData {
+    init_data: Vec<u8>,
+    /// A map of tracked data indices to the offset in `init_data` where that
+    /// particular value begins. The offsets are in ascending order. The indices
+    /// are unordered.
+    init_offsets: Vec<(u8, u32)>,
+    update_data: Vec<u8>,
+}
+
+impl TrackedData {
+    /// Returns initial tracked data for the entity, ready to be sent in the
+    /// [`EntityTrackerUpdateS2c`][packet] packet. This is used when the entity
+    /// enters the view of a client.
+    ///
+    /// [packet]: valence_protocol::packet::s2c::play::EntityTrackerUpdateS2c
+    pub fn init_data(&self) -> Option<&[u8]> {
+        if self.init_data.len() > 1 {
+            Some(&self.init_data)
+        } else {
+            None
+        }
+    }
+
+    /// Contains updated tracked data for the entity, ready to be sent in the
+    /// [`EntityTrackerUpdateS2c`][packet] packet. This is used when tracked
+    /// data is changed and the client is already in view of the entity.
+    ///
+    /// [packet]: valence_protocol::packet::s2c::play::EntityTrackerUpdateS2c
+    pub fn update_data(&self) -> Option<&[u8]> {
+        if self.update_data.len() > 1 {
+            Some(&self.update_data)
+        } else {
+            None
+        }
+    }
+
+    pub fn append_init_value(&mut self, index: u8, type_id: u8, value: &impl Encode) {
+        debug_assert!(
+            index != 0xff,
+            "index of 0xff is reserved for the terminator"
+        );
+
+        self.init_data.pop(); // Remove terminator.
+
+        // Append the new value to the end.
+        debug_assert!(u32::try_from(self.init_data.len()).is_ok(), "too much data");
+        let new_offset = self.init_data.len() as u32;
+
+        self.init_data.extend_from_slice(&[index, type_id]);
+        if let Err(e) = value.encode(&mut self.init_data) {
+            warn!("failed to encode initial tracked data: {e:#}");
+        }
+
+        self.init_offsets.push((index, new_offset));
+
+        self.init_data.push(0xff); // Add terminator.
+    }
+
+    pub fn remove_init_value(&mut self, index: u8) {
+        // Is the index in the data?
+        if let Some((offset_pos, (_, start))) = self
+            .init_offsets
+            .iter()
+            .enumerate()
+            .find(|(_, (idx, _))| *idx == index)
+        {
+            let start = *start as usize;
+
+            let end = self
+                .init_offsets
+                .get(offset_pos + 1)
+                .map(|(_, end)| *end as usize)
+                .unwrap_or(self.init_data.len() - 1); // -1 to skip terminator.
+
+            // Remove the range of bytes for the value.
+            self.init_data.drain(start..end);
+            self.init_offsets.remove(offset_pos);
+        }
+    }
+
+    pub fn append_update_value(&mut self, index: u8, type_id: u8, value: &impl Encode) {
+        debug_assert!(
+            index != 0xff,
+            "index of 0xff is reserved for the terminator"
+        );
+
+        self.update_data.pop(); // Remove terminator.
+
+        self.update_data.extend_from_slice(&[index, type_id]);
+        if let Err(e) = value.encode(&mut self.update_data) {
+            warn!("failed to encode updated tracked data: {e:#}");
+        }
+
+        self.update_data.push(0xff); // Add terminator.
+    }
+
+    pub fn clear_update_values(&mut self) {
+        self.update_data.clear();
+    }
 }
 
 #[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Debug, Encode, Decode)]
@@ -271,7 +370,7 @@ pub struct EulerAngle {
 }
 
 #[derive(Copy, Clone)]
-pub struct OptionalInt(pub Option<i32>);
+struct OptionalInt(Option<i32>);
 
 impl Encode for OptionalInt {
     fn encode(&self, w: impl std::io::Write) -> anyhow::Result<()> {
@@ -287,7 +386,7 @@ impl Encode for OptionalInt {
 impl Decode<'_> for OptionalInt {
     fn decode(r: &mut &[u8]) -> anyhow::Result<Self> {
         let n = VarInt::decode(r)?.0;
-        
+
         Ok(Self(if n == 0 {
             None
         } else {
@@ -297,7 +396,7 @@ impl Decode<'_> for OptionalInt {
 }
 
 /// Maintains information about all spawned Minecraft entities.
-#[derive(Resource)]
+#[derive(Resource, Debug)]
 pub struct EntityManager {
     /// Maps protocol IDs to ECS entities.
     id_to_entity: FxHashMap<i32, Entity>,
@@ -364,16 +463,12 @@ impl Plugin for EntityPlugin {
 }
 
 fn init_entities(
-    mut entities: Query<(Entity, &mut EntityId, &mut UniqueId), Added<BaseEntity>>,
+    mut entities: Query<(Entity, &mut EntityId, &mut UniqueId), Added<EntityKind>>,
     mut manager: ResMut<EntityManager>,
 ) {
-    for (entity, mut id, mut uuid) in &mut entities {
-        if id.0 == 0 {
+    for (entity, mut id, uuid) in &mut entities {
+        if *id == EntityId::default() {
             *id = manager.next_id();
-        }
-
-        if uuid.0.is_nil() {
-            uuid.0 = Uuid::from_bytes(manager.uuid_rng.gen());
         }
 
         if let Some(conflict) = manager.id_to_entity.insert(id.0, entity) {
@@ -393,10 +488,10 @@ fn init_entities(
 }
 
 fn remove_despawned_from_manager(
-    entities: Query<(Entity, &EntityId, &UniqueId), (With<BaseEntity>, With<Despawned>)>,
+    entities: Query<(&EntityId, &UniqueId), (With<EntityKind>, With<Despawned>)>,
     mut manager: ResMut<EntityManager>,
 ) {
-    for (entity, id, uuid) in &entities {
+    for (id, uuid) in &entities {
         manager.id_to_entity.remove(&id.0);
         manager.uuid_to_entity.remove(&uuid.0);
     }
